requirements.txt
aniso8601==9.0.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.10.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.5
capstone==5.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
colored-traceback==0.3.0
cryptography==38.0.4
dbus-python==1.3.2
debtags==2.1
Deprecated==1.2.14
dicttoxml==1.7.15
distro==1.8.0
distro-info==1.6
dnspython==2.4.2
flasgger==0.9.5
Flask==2.2.5
Flask-RESTful==0.3.9
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.20.4
httpx==0.23.3
hyperframe==6.0.0
idna==3.4
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.10.3
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lightdm-gtk-greeter-settings==1.2.2
louis==3.27.0
lxml==4.9.3
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mechanize==0.4.8
mistune0==0.8.4
netaddr==0.8.0
oauthlib==3.2.2
onboard==1.4.1
packaging==23.1
paramiko==3.3.1
plumbum==1.8.2
psutil==5.9.5
pwn==1.0
pwntools==4.10.0
pycairo==1.24.0
pycparser==2.21
pyelftools==0.30
Pygments==2.16.1
PyGObject==3.46.0
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0.1
redis==4.3.4
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rfc3987==1.3.8
rich==13.3.1
ROPGadget==7.4
rpyc==5.3.1
rq==1.15.0
simplejson==3.19.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.5
tk==0.1.0
unicodecsv==0.14.1
unicorn==2.0.1.post1
uritemplate==4.1.1
urllib3==1.26.18
wadllib==1.3.6
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.15.0
xdg==5
XlsxWriter==3.0.2
